documentation/src/requirements.txt
Pillow==9.0.0
numpy==1.21.2
matplotlib==3.4.3
tensorflow==2.8.0
scikit-learn==1.0.1
seaborn==0.11.2